refactor: reorder lib file structure as described in #48
