+++ lib/model/settings/status.dart
