+++ lib/model/trupp/trupp.dart
