+++ lib/model/trupp/trupp.freezed.dart
