+++ lib/repositories/firefighters_repository.dart
@@ -4,7 +4,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import '../firebase/firestore_service.dart';
 import '../firebase/firestore_provider.dart';
 import '../firebase/firebase_auth_provider.dart';
-import '../ui/model/settings/firefighter.dart';
+import '../model/settings/firefighter.dart';
 
 // Repository class for managing Firefighter data in Firestore.
 class FirefightersRepository {
