+++ lib/repositories/initial_settings_repository.dart
@@ -4,7 +4,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import '../firebase/firestore_service.dart';
 import '../firebase/firestore_provider.dart';
 import '../firebase/firebase_auth_provider.dart';
-import '../ui/model/settings/initial_settings.dart';
+import '../model/settings/initial_settings.dart';
 
 // Repository class for managing Initial Settings data in Firestore.
 // Note: Initial settings are stored as a single document, not a collection.
