+++ lib/ui/app.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'router/router.dart';
+import '../router/router.dart';
 
 class AsuApp extends ConsumerWidget {
   const AsuApp({super.key});
